:art: WineEmbeddingUtility rename 및 Client에서 OpenAI 관련없으면 분리
